pyproject.toml
[project]
name = "vestigium_foundry_spells"
version = "0.1.0"
description = ""
authors = [
    {name = "Rocko",email = "vbernier@nexapp.ca"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "notebook (>=7.4.0,<8.0.0)",
    "inflection (>=0.5.1,<0.6.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "image (>=1.5.33,<2.0.0)",
    "diffusers (>=0.33.0,<0.34.0)",
    "torch (>=2.6.0,<3.0.0)",
    "transformers (>=4.51.1,<5.0.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "hf-xet (>=1.0.2,<2.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "weasyprint (>=65.0,<66.0)",
    "ruff (>=0.11.5,<0.12.0)",
    "pyright (>=1.1.399,<2.0.0)",
    "google-api-python-client (>=2.167.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pypdf2 = "^3.0.1"

